+++ src/Geta.Optimizely.Sitemaps/Utils/IContentFilter.cs
@@ -2,7 +2,6 @@
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
 using EPiServer.Core;
-using EPiServer.Web;
 using Geta.Optimizely.Sitemaps.Entities;
 
 namespace Geta.Optimizely.Sitemaps.Utils
@@ -10,7 +9,6 @@ namespace Geta.Optimizely.Sitemaps.Utils
     public interface IContentFilter
     {
         bool ShouldExcludeContent(IContent content);
-        bool ShouldExcludeContent(
-            CurrentLanguageContent languageContentInfo, SiteDefinition siteSettings, SitemapData sitemapData);
+        bool ShouldExcludeContent(CurrentLanguageContent languageContentInfo, SitemapData sitemapData);
     }
 }