+++ src/Geta.Optimizely.Sitemaps/XML/StandardSitemapXmlGenerator.cs
@@ -2,9 +2,9 @@
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
 using EPiServer;
+using EPiServer.Applications;
 using EPiServer.DataAbstraction;
 using EPiServer.Framework.Cache;
-using EPiServer.Web;
 using EPiServer.Web.Routing;
 using Geta.Optimizely.Sitemaps.Repositories;
 using Geta.Optimizely.Sitemaps.Services;
@@ -20,7 +20,7 @@ namespace Geta.Optimizely.Sitemaps.XML
             ISitemapRepository sitemapRepository,
             IContentRepository contentRepository,
             IUrlResolver urlResolver,
-            ISiteDefinitionRepository siteDefinitionRepository,
+            IApplicationRepository applicationRepository,
             ILanguageBranchRepository languageBranchRepository,
             IContentFilter contentFilter,
             IUriAugmenterService uriAugmenterService,
@@ -31,7 +31,7 @@ namespace Geta.Optimizely.Sitemaps.XML
                 sitemapRepository,
                 contentRepository,
                 urlResolver,
-                siteDefinitionRepository,
+                applicationRepository,
                 languageBranchRepository,
                 contentFilter,
                 uriAugmenterService,